requirements.txt
pandas==1.5.2
lxml==4.9.2
request==2.28.1
python-dotenv~=1.0.0
progressbar2==3.53.3
deepdiff==6.3.1